+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 - Improved instructions for developers
 - Added link to ssa.tools on `Fixed Income` page
+- Fixed bug on max age range for SS when month != 0
 
 ### Version 2025.12.05
 - Added instructions for obtaining PIA